Move nfc_tag_type_t to NFCDefinitions.h
